requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.4
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
Pillow==10.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.18
SQLAlchemy-Utils==0.41.1
typing_extensions==4.7.1
Werkzeug==2.3.6
